# docs/swagger.yaml
definitions:
  input.CreateAccountInputVO:
    properties:
      balance:
        type: integer
      cpf:
        type: string
      name:
        type: string
      secret:
        type: string
    type: object
  input.CreateTransferInputVO:
    properties:
      account_destination_id:
        type: string
      accountOriginID:
        type: string
      amount:
        type: integer
    type: object
  input.CredentialInputVO:
    properties:
      cpf:
        type: string
      secret:
        type: string
    type: object
  output.CreateAccountOutputVO:
    properties:
      balance:
        type: number
      cpf:
        type: string
      created_at:
        type: string
      name:
        type: string
    type: object
  output.CreateTransferOutputVO:
    properties:
      account_destination_id:
        type: string
      account_origin_id:
        type: string
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
    type: object
  output.FindAccountBalanceOutputVO:
    properties:
      balance:
        type: number
      id:
        type: string
    type: object
  output.FindAllAccountOutputVO:
    properties:
      balance:
        type: number
      cpf:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  output.FindAllTransferOutputVO:
    properties:
      account_destination_id:
        type: string
      account_origin_id:
        type: string
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      message:
        example: internal server error
        type: string
    type: object
  response.HTTPErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorResponse'
      statusCode:
        example: 500
        type: integer
    type: object
info:
  contact:
    email: brunnodecastro@gmail.com
    name: Bruno de Castro Oliveira
  title: Digital Accounts API
  version: "1.0"
paths:
  /account/{account_id}/balance:
    get:
      consumes:
      - application/json
      description: Gets the account balance
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/output.FindAccountBalanceOutputVO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
        "500":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
      summary: Gets the account balance
      tags:
      - Accounts
  /accounts:
    get:
      description: List all accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/output.FindAllAccountOutputVO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
      summary: List all accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Creates a new account
      parameters:
      - description: Account Input
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/input.CreateAccountInputVO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/output.CreateAccountOutputVO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
        "500":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
      summary: Creates a new account
      tags:
      - Accounts
  /login:
    post:
      consumes:
      - application/json
      description: authenticate the user in the api
      parameters:
      - description: Credential Input
        in: body
        name: credential
        required: true
        schema:
          $ref: '#/definitions/input.CredentialInputVO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/output.CreateTransferOutputVO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: authenticate the user
      tags:
      - Authentication
  /transfers:
    get:
      description: List all transfers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/output.FindAllTransferOutputVO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all transfers
      tags:
      - Transfers
    post:
      consumes:
      - application/json
      description: Creates a new transfer
      parameters:
      - description: Transfer Input
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/input.CreateTransferInputVO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/output.CreateTransferOutputVO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
        "500":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Creates a new transfer
      tags:
      - Transfers
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"